requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
influxdb==5.3.2
msgpack==1.0.8
mysqlclient==2.2.4
paho-mqtt==2.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
urllib3==2.2.1
